fix: resolve integration problems with dumpster-microservice

+++ app/src/main/kotlin/swc/microservice/mission/entities/Position.kt
@@ -1,3 +1,3 @@
 package swc.microservice.mission.entities
 
-data class Position(val latitude: Long, val longitude: Long)
+data class Position(val latitude: Double, val longitude: Double)
